refactor: constimpl pacakge
