+++ kotlin-utils/src/main/kotlin/xyz/lbres/kotlinutils/set/multiset/impl/AbstractMultiSet.kt
@@ -7,9 +7,9 @@ import xyz.lbres.kotlinutils.set.multiset.MultiSet
 import kotlin.math.min
 
 /**
- * Partial MultiSet implementation with functionality that can be shared by [StandardMultiSet] and [MutableStandardMultiSet].
+ * Partial MultiSet implementation with functionality that can be shared by [MultiSetImpl] and [MutableMultiSetImpl].
  */
-internal abstract class AbstractStandardMultiSet<E> : MultiSet<E> {
+internal abstract class AbstractMultiSet<E> : MultiSet<E> {
     /**
      * All distinct values contained in the MultiSet.
      */
@@ -120,7 +120,7 @@ internal abstract class AbstractStandardMultiSet<E> : MultiSet<E> {
     private fun genericBinaryOperation(other: MultiSet<E>, operation: (count: Int, otherCount: Int) -> Int, useAllValues: Boolean = true): MultiSet<E> {
         val counts = getCounts()
 
-        val newCounts: Map<E, Int> = if (other is AbstractStandardMultiSet<E>) {
+        val newCounts: Map<E, Int> = if (other is AbstractMultiSet<E>) {
             val otherCounts = other.getCounts()
             val allValues = simpleIf(useAllValues, { counts.keys + otherCounts.keys }, { counts.keys })
