+++ kotlin-utils/src/main/kotlin/xyz/lbres/kotlinutils/set/multiset/impl/MutableMultiSetImpl.kt
@@ -7,7 +7,7 @@ import kotlin.math.min
 /**
  * Mutable set implementation that allows multiple occurrences of the same value.
  */
-internal class MutableStandardMultiSet<E> : AbstractStandardMultiSet<E>, MutableMultiSet<E> {
+internal class MutableMultiSetImpl<E> : AbstractMultiSet<E>, MutableMultiSet<E> {
     /**
      * Number of elements in set.
      */
@@ -138,7 +138,7 @@ internal class MutableStandardMultiSet<E> : AbstractStandardMultiSet<E>, Mutable
      * @return [MutableMultiSet]<E>: MultiSet containing the items in this MultiSet but not the other
      */
     override operator fun minus(other: MultiSet<E>): MutableMultiSet<E> {
-        return super<AbstractStandardMultiSet>.minus(other) as MutableMultiSet<E>
+        return super<AbstractMultiSet>.minus(other) as MutableMultiSet<E>
     }
 
     /**
@@ -149,13 +149,13 @@ internal class MutableStandardMultiSet<E> : AbstractStandardMultiSet<E>, Mutable
      * @return [MutableMultiSet]<E>: MultiSet containing all values from both MultiSets
      */
     override operator fun plus(other: MultiSet<E>): MutableMultiSet<E> {
-        return super<AbstractStandardMultiSet>.plus(other) as MutableMultiSet<E>
+        return super<AbstractMultiSet>.plus(other) as MutableMultiSet<E>
     }
 
     /**
      * Initialize a new MutableMultiSet from existing counts.
      */
-    override fun createFromCounts(counts: Map<E, Int>): MultiSet<E> = MutableStandardMultiSet(counts)
+    override fun createFromCounts(counts: Map<E, Int>): MultiSet<E> = MutableMultiSetImpl(counts)
 
     /**
      * Get an iterator for the elements in this set.
