+++ kotlin-utils/src/test/kotlin/xyz/lbres/kotlinutils/set/multiset/MultiSetUtilsTest.kt
@@ -1,9 +1,9 @@
 package xyz.lbres.kotlinutils.set.multiset
 
-import xyz.lbres.kotlinutils.set.multiset.impl.ConstMultiSetImpl
-import xyz.lbres.kotlinutils.set.multiset.impl.MutableConstMultiSetImpl
-import xyz.lbres.kotlinutils.set.multiset.impl.MutableStandardMultiSet
-import xyz.lbres.kotlinutils.set.multiset.impl.StandardMultiSet
+import xyz.lbres.kotlinutils.set.multiset.impl.constimpl.ConstMultiSetImpl
+import xyz.lbres.kotlinutils.set.multiset.impl.MultiSetImpl
+import xyz.lbres.kotlinutils.set.multiset.impl.constimpl.ConstMutableMultiSetImpl
+import xyz.lbres.kotlinutils.set.multiset.impl.MutableMultiSetImpl
 import kotlin.test.Test
 import kotlin.test.assertEquals
 
@@ -11,62 +11,62 @@ class MultiSetUtilsTest {
     @Test
     fun testMultiSetOf() {
         var set: MultiSet<Int> = multiSetOf()
-        var expected: MultiSet<Int> = StandardMultiSet(emptyList())
+        var expected: MultiSet<Int> = MultiSetImpl(emptyList())
         assertEquals(expected, set)
 
         set = multiSetOf(1)
-        expected = StandardMultiSet(listOf(1))
+        expected = MultiSetImpl(listOf(1))
         assertEquals(expected, set)
 
         set = multiSetOf(1, 2, 3, 4)
-        expected = StandardMultiSet(listOf(1, 2, 3, 4))
+        expected = MultiSetImpl(listOf(1, 2, 3, 4))
         assertEquals(expected, set)
 
         set = multiSetOf(1, 1, 1)
-        expected = StandardMultiSet(listOf(1, 1, 1))
+        expected = MultiSetImpl(listOf(1, 1, 1))
         assertEquals(expected, set)
 
         val stringSet = multiSetOf("", "hello", "world")
-        val stringExpected = StandardMultiSet(listOf("", "hello", "world"))
+        val stringExpected = MultiSetImpl(listOf("", "hello", "world"))
         assertEquals(stringExpected, stringSet)
 
         val listSet = multiSetOf(listOf(123), listOf(1, 4, 5, 6), listOf(99, 100, 97))
-        val listExpected = StandardMultiSet(listOf(listOf(123), listOf(1, 4, 5, 6), listOf(99, 100, 97)))
+        val listExpected = MultiSetImpl(listOf(listOf(123), listOf(1, 4, 5, 6), listOf(99, 100, 97)))
         assertEquals(listExpected, listSet)
 
         val compListSet = multiSetOf(listOf(1, 2, 3), listOf("abc", "def"), listOf("abc", "def"))
-        val compListSetExpected = StandardMultiSet(listOf(listOf(1, 2, 3), listOf("abc", "def"), listOf("abc", "def")))
+        val compListSetExpected = MultiSetImpl(listOf(listOf(1, 2, 3), listOf("abc", "def"), listOf("abc", "def")))
         assertEquals(compListSetExpected, compListSet)
     }
 
     @Test
     fun testMutableMultiSetOf() {
         var set: MutableMultiSet<Int> = mutableMultiSetOf()
-        var expected: MutableMultiSet<Int> = MutableStandardMultiSet(emptyList())
+        var expected: MutableMultiSet<Int> = MutableMultiSetImpl(emptyList())
         assertEquals(expected, set)
 
         set = mutableMultiSetOf(1)
-        expected = MutableStandardMultiSet(listOf(1))
+        expected = MutableMultiSetImpl(listOf(1))
         assertEquals(expected, set)
 
         set = mutableMultiSetOf(1, 2, 3, 4)
-        expected = MutableStandardMultiSet(listOf(1, 2, 3, 4))
+        expected = MutableMultiSetImpl(listOf(1, 2, 3, 4))
         assertEquals(expected, set)
 
         set = mutableMultiSetOf(1, 1, 1)
-        expected = MutableStandardMultiSet(listOf(1, 1, 1))
+        expected = MutableMultiSetImpl(listOf(1, 1, 1))
         assertEquals(expected, set)
 
         val stringSet = mutableMultiSetOf("", "hello", "world")
-        val stringExpected = MutableStandardMultiSet(listOf("", "hello", "world"))
+        val stringExpected = MutableMultiSetImpl(listOf("", "hello", "world"))
         assertEquals(stringExpected, stringSet)
 
         val listSet = mutableMultiSetOf(listOf(123), listOf(1, 4, 5, 6), listOf(99, 100, 97))
-        val listExpected = MutableStandardMultiSet(listOf(listOf(123), listOf(1, 4, 5, 6), listOf(99, 100, 97)))
+        val listExpected = MutableMultiSetImpl(listOf(listOf(123), listOf(1, 4, 5, 6), listOf(99, 100, 97)))
         assertEquals(listExpected, listSet)
 
         val compListSet = mutableMultiSetOf(listOf(1, 2, 3), listOf("abc", "def"), listOf("abc", "def"))
-        val compListSetExpected = MutableStandardMultiSet(listOf(listOf(1, 2, 3), listOf("abc", "def"), listOf("abc", "def")))
+        val compListSetExpected = MutableMultiSetImpl(listOf(listOf(1, 2, 3), listOf("abc", "def"), listOf("abc", "def")))
         assertEquals(compListSetExpected, compListSet)
     }
 
@@ -104,31 +104,31 @@ class MultiSetUtilsTest {
     @Test
     fun testMutableConstMultiSetOf() {
         var set: MutableMultiSet<Int> = constMutableMultiSetOf()
-        var expected: MutableMultiSet<Int> = MutableConstMultiSetImpl(emptyList())
+        var expected: MutableMultiSet<Int> = ConstMutableMultiSetImpl(emptyList())
         assertEquals(expected, set)
 
         set = constMutableMultiSetOf(1)
-        expected = MutableConstMultiSetImpl(listOf(1))
+        expected = ConstMutableMultiSetImpl(listOf(1))
         assertEquals(expected, set)
 
         set = constMutableMultiSetOf(1, 2, 3, 4)
-        expected = MutableConstMultiSetImpl(listOf(1, 2, 3, 4))
+        expected = ConstMutableMultiSetImpl(listOf(1, 2, 3, 4))
         assertEquals(expected, set)
 
         set = constMutableMultiSetOf(1, 1, 1)
-        expected = MutableConstMultiSetImpl(listOf(1, 1, 1))
+        expected = ConstMutableMultiSetImpl(listOf(1, 1, 1))
         assertEquals(expected, set)
 
         val stringSet = constMutableMultiSetOf("", "hello", "world")
-        val stringExpected = MutableConstMultiSetImpl(listOf("", "hello", "world"))
+        val stringExpected = ConstMutableMultiSetImpl(listOf("", "hello", "world"))
         assertEquals(stringExpected, stringSet)
 
         val listSet = constMutableMultiSetOf(listOf(123), listOf(1, 4, 5, 6), listOf(99, 100, 97))
-        val listExpected = MutableConstMultiSetImpl(listOf(listOf(123), listOf(1, 4, 5, 6), listOf(99, 100, 97)))
+        val listExpected = ConstMutableMultiSetImpl(listOf(listOf(123), listOf(1, 4, 5, 6), listOf(99, 100, 97)))
         assertEquals(listExpected, listSet)
 
         val compListSet = constMutableMultiSetOf(listOf(1, 2, 3), listOf("abc", "def"), listOf("abc", "def"))
-        val compListSetExpected = MutableConstMultiSetImpl(listOf(listOf(1, 2, 3), listOf("abc", "def"), listOf("abc", "def")))
+        val compListSetExpected = ConstMutableMultiSetImpl(listOf(listOf(1, 2, 3), listOf("abc", "def"), listOf("abc", "def")))
         assertEquals(compListSetExpected, compListSet)
     }
 
@@ -156,20 +156,20 @@ class MultiSetUtilsTest {
 
         set = MultiSet(5) { 2 }
         expectedSize = 5
-        expected = StandardMultiSet(listOf(2, 2, 2, 2, 2))
+        expected = MultiSetImpl(listOf(2, 2, 2, 2, 2))
         assertEquals(expectedSize, set.size)
         assertEquals(expected, set)
 
         set = MultiSet(4) { 3 * it }
         expectedSize = 4
-        expected = StandardMultiSet(listOf(0, 3, 6, 9))
+        expected = MultiSetImpl(listOf(0, 3, 6, 9))
         assertEquals(expectedSize, set.size)
         assertEquals(expected, set)
 
         val values = listOf(4, 6, 7, 8, 9, 11, -3, -3)
         set = MultiSet(8) { values[it] }
         expectedSize = 8
-        expected = StandardMultiSet(values)
+        expected = MultiSetImpl(values)
         assertEquals(expectedSize, set.size)
         assertEquals(expected, set)
     }
@@ -184,20 +184,20 @@ class MultiSetUtilsTest {
 
         set = MutableMultiSet(5) { 2 }
         expectedSize = 5
-        expected = MutableStandardMultiSet(listOf(2, 2, 2, 2, 2))
+        expected = MutableMultiSetImpl(listOf(2, 2, 2, 2, 2))
         assertEquals(expectedSize, set.size)
         assertEquals(expected, set)
 
         set = MutableMultiSet(4) { 3 * it }
         expectedSize = 4
-        expected = MutableStandardMultiSet(listOf(0, 3, 6, 9))
+        expected = MutableMultiSetImpl(listOf(0, 3, 6, 9))
         assertEquals(expectedSize, set.size)
         assertEquals(expected, set)
 
         val values = listOf(4, 6, 7, 8, 9, 11, -3, -3)
         set = MutableMultiSet(8) { values[it] }
         expectedSize = 8
-        expected = MutableStandardMultiSet(values)
+        expected = MutableMultiSetImpl(values)
         assertEquals(expectedSize, set.size)
         assertEquals(expected, set)
     }
