+++ kotlin-utils/src/test/kotlin/xyz/lbres/kotlinutils/set/multiset/impl/MutableMultiSetImplTest.kt
@@ -2,17 +2,17 @@ package xyz.lbres.kotlinutils.set.multiset.impl
 
 import xyz.lbres.kotlinutils.list.IntList
 import xyz.lbres.kotlinutils.set.multiset.* // ktlint-disable no-wildcard-imports no-unused-imports
-import xyz.lbres.kotlinutils.set.multiset.impl.mutableconst.* // ktlint-disable no-wildcard-imports no-unused-imports
+import xyz.lbres.kotlinutils.set.multiset.impl.mutable.* // ktlint-disable no-wildcard-imports no-unused-imports
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertFalse
 
-class MutableConstMultiSetImplImplTest {
-    @Test fun testConstructor() = runMutableConstConstructorTests()
-    @Test fun testEquals() = runMutableConstEqualsTests()
+class MutableMultiSetImplTest {
+    @Test fun testConstructor() = runMutableConstructorTests()
+    @Test fun testEquals() = runMutableEqualsTests()
 
-    @Test fun testContains() = runMutableConstContainsTests()
-    @Test fun testContainsAll() = runMutableConstContainsAllTests()
+    @Test fun testContains() = runMutableContainsTests()
+    @Test fun testContainsAll() = runMutableContainsAllTests()
 
     @Test fun testClear() = runClearTests()
     @Test fun testAdd() = runAddTests()
@@ -21,20 +21,20 @@ class MutableConstMultiSetImplImplTest {
     @Test fun testRemoveAll() = runRemoveAllTests()
     @Test fun testRetainAll() = runRetainAllTests()
 
-    @Test fun testMinus() = runMutableConstMinusTests()
-    @Test fun testPlus() = runMutableConstPlusTests()
-    @Test fun testIntersect() = runMutableConstIntersectTests()
+    @Test fun testMinus() = runMutableMinusTests()
+    @Test fun testPlus() = runMutablePlusTests()
+    @Test fun testIntersect() = runMutableIntersectTests()
 
-    @Test fun testIsEmpty() = runMutableConstIsEmptyTests()
-    @Test fun testGetCountOf() = runMutableConstGetCountOfTests()
+    @Test fun testIsEmpty() = runMutableIsEmptyTests()
+    @Test fun testGetCountOf() = runMutableGetCountOfTests()
 
     @Test
     fun testIterator() {
-        var set: ConstMutableMultiSet<Int> = constMutableMultiSetOf()
+        var set: MutableMultiSet<Int> = mutableMultiSetOf()
         var iter = set.iterator()
         assertFalse(iter.hasNext())
 
-        set = constMutableMultiSetOf(1, 2, 3, 4)
+        set = mutableMultiSetOf(1, 2, 3, 4)
         iter = set.iterator()
         var values: MutableList<Int> = mutableListOf()
         var expected = listOf(1, 2, 3, 4)
@@ -43,7 +43,7 @@ class MutableConstMultiSetImplImplTest {
         }
         assertEquals(expected.sorted(), values.sorted())
 
-        set = constMutableMultiSetOf(1, 2, 3, 4, 1, 4, 5)
+        set = mutableMultiSetOf(1, 2, 3, 4, 1, 4, 5)
         iter = set.iterator()
         values = mutableListOf()
         expected = listOf(1, 1, 2, 3, 4, 4, 5)
@@ -72,7 +72,7 @@ class MutableConstMultiSetImplImplTest {
 
         val mutableList1 = mutableListOf(1, 2, 3)
         val mutableList2 = mutableListOf(0, 5, 7)
-        val listSet: ConstMutableMultiSet<IntList> = constMutableMultiSetOf(mutableList1, mutableList2)
+        val listSet: MutableMultiSet<IntList> = mutableMultiSetOf(mutableList1, mutableList2)
 
         var listIter = listSet.iterator()
         var listExpected = listOf(listOf(1, 2, 3), listOf(0, 5, 7))
@@ -94,35 +94,39 @@ class MutableConstMultiSetImplImplTest {
 
     @Test
     fun testToString() {
-        var set: ConstMutableMultiSet<Int> = constMutableMultiSetOf()
+        var set: MutableMultiSet<Int> = mutableMultiSetOf()
         var expected = "[]"
         assertEquals(expected, set.toString())
 
-        set = constMutableMultiSetOf(100, -200, 400, 800)
+        set = mutableMultiSetOf(100, -200, 400, 800)
         expected = "[100, -200, 400, 800]"
         assertEquals(expected, set.toString())
 
-        set = constMutableMultiSetOf(1, 1, 1, 1)
+        set = mutableMultiSetOf(1, 1, 1, 1)
         expected = "[1, 1, 1, 1]"
         assertEquals(expected, set.toString())
 
-        set = constMutableMultiSetOf(2, 4, 2, 1)
-        expected = "[2, 2, 4, 1]"
+        set = mutableMultiSetOf(2, 4, 2, 1)
+        expected = "[2, 4, 2, 1]"
         assertEquals(expected, set.toString())
 
         set.add(5)
-        expected = "[2, 2, 4, 1, 5]"
+        expected = "[2, 4, 2, 1, 5]"
         assertEquals(expected, set.toString())
 
         set.remove(2)
-        expected = "[2, 4, 1, 5]"
+        expected = "[4, 2, 1, 5]"
         assertEquals(expected, set.toString())
 
-        val list1 = listOf(1, 2, 3)
-        val list2 = listOf(0, 5, 7)
-        val listSet: ConstMutableMultiSet<IntList> = constMutableMultiSetOf(list1, list2)
+        val mutableList1 = mutableListOf(1, 2, 3)
+        val mutableList2 = mutableListOf(0, 5, 7)
+        val listSet: MutableMultiSet<IntList> = mutableMultiSetOf(mutableList1, mutableList2)
 
         expected = "[[1, 2, 3], [0, 5, 7]]"
+        assertEquals(expected, listSet.toString())
+
+        mutableList2.clear()
+        expected = "[[1, 2, 3], []]"
         assertEquals(expected, listSet.toString())
     }
 }
