+++ kotlin-utils/src/test/kotlin/xyz/lbres/kotlinutils/set/multiset/impl/constimpl/ConstMutableMultiSetImplImplTest.kt
@@ -1,18 +1,18 @@
-package xyz.lbres.kotlinutils.set.multiset.impl
+package xyz.lbres.kotlinutils.set.multiset.impl.constimpl
 
 import xyz.lbres.kotlinutils.list.IntList
 import xyz.lbres.kotlinutils.set.multiset.* // ktlint-disable no-wildcard-imports no-unused-imports
-import xyz.lbres.kotlinutils.set.multiset.impl.mutable.* // ktlint-disable no-wildcard-imports no-unused-imports
+import xyz.lbres.kotlinutils.set.multiset.impl.constimpl.mutable.*
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertFalse
 
-class MutableStandardMultiSetTest {
-    @Test fun testConstructor() = runMutableConstructorTests()
-    @Test fun testEquals() = runMutableEqualsTests()
+class ConstMutableMultiSetImplImplTest {
+    @Test fun testConstructor() = runMutableConstConstructorTests()
+    @Test fun testEquals() = runMutableConstEqualsTests()
 
-    @Test fun testContains() = runMutableContainsTests()
-    @Test fun testContainsAll() = runMutableContainsAllTests()
+    @Test fun testContains() = runMutableConstContainsTests()
+    @Test fun testContainsAll() = runMutableConstContainsAllTests()
 
     @Test fun testClear() = runClearTests()
     @Test fun testAdd() = runAddTests()
@@ -21,20 +21,20 @@ class MutableStandardMultiSetTest {
     @Test fun testRemoveAll() = runRemoveAllTests()
     @Test fun testRetainAll() = runRetainAllTests()
 
-    @Test fun testMinus() = runMutableMinusTests()
-    @Test fun testPlus() = runMutablePlusTests()
-    @Test fun testIntersect() = runMutableIntersectTests()
+    @Test fun testMinus() = runMutableConstMinusTests()
+    @Test fun testPlus() = runMutableConstPlusTests()
+    @Test fun testIntersect() = runMutableConstIntersectTests()
 
-    @Test fun testIsEmpty() = runMutableIsEmptyTests()
-    @Test fun testGetCountOf() = runMutableGetCountOfTests()
+    @Test fun testIsEmpty() = runMutableConstIsEmptyTests()
+    @Test fun testGetCountOf() = runMutableConstGetCountOfTests()
 
     @Test
     fun testIterator() {
-        var set: MutableMultiSet<Int> = mutableMultiSetOf()
+        var set: ConstMutableMultiSet<Int> = constMutableMultiSetOf()
         var iter = set.iterator()
         assertFalse(iter.hasNext())
 
-        set = mutableMultiSetOf(1, 2, 3, 4)
+        set = constMutableMultiSetOf(1, 2, 3, 4)
         iter = set.iterator()
         var values: MutableList<Int> = mutableListOf()
         var expected = listOf(1, 2, 3, 4)
@@ -43,7 +43,7 @@ class MutableStandardMultiSetTest {
         }
         assertEquals(expected.sorted(), values.sorted())
 
-        set = mutableMultiSetOf(1, 2, 3, 4, 1, 4, 5)
+        set = constMutableMultiSetOf(1, 2, 3, 4, 1, 4, 5)
         iter = set.iterator()
         values = mutableListOf()
         expected = listOf(1, 1, 2, 3, 4, 4, 5)
@@ -72,7 +72,7 @@ class MutableStandardMultiSetTest {
 
         val mutableList1 = mutableListOf(1, 2, 3)
         val mutableList2 = mutableListOf(0, 5, 7)
-        val listSet: MutableMultiSet<IntList> = mutableMultiSetOf(mutableList1, mutableList2)
+        val listSet: ConstMutableMultiSet<IntList> = constMutableMultiSetOf(mutableList1, mutableList2)
 
         var listIter = listSet.iterator()
         var listExpected = listOf(listOf(1, 2, 3), listOf(0, 5, 7))
@@ -94,39 +94,35 @@ class MutableStandardMultiSetTest {
 
     @Test
     fun testToString() {
-        var set: MutableMultiSet<Int> = mutableMultiSetOf()
+        var set: ConstMutableMultiSet<Int> = constMutableMultiSetOf()
         var expected = "[]"
         assertEquals(expected, set.toString())
 
-        set = mutableMultiSetOf(100, -200, 400, 800)
+        set = constMutableMultiSetOf(100, -200, 400, 800)
         expected = "[100, -200, 400, 800]"
         assertEquals(expected, set.toString())
 
-        set = mutableMultiSetOf(1, 1, 1, 1)
+        set = constMutableMultiSetOf(1, 1, 1, 1)
         expected = "[1, 1, 1, 1]"
         assertEquals(expected, set.toString())
 
-        set = mutableMultiSetOf(2, 4, 2, 1)
-        expected = "[2, 4, 2, 1]"
+        set = constMutableMultiSetOf(2, 4, 2, 1)
+        expected = "[2, 2, 4, 1]"
         assertEquals(expected, set.toString())
 
         set.add(5)
-        expected = "[2, 4, 2, 1, 5]"
+        expected = "[2, 2, 4, 1, 5]"
         assertEquals(expected, set.toString())
 
         set.remove(2)
-        expected = "[4, 2, 1, 5]"
+        expected = "[2, 4, 1, 5]"
         assertEquals(expected, set.toString())
 
-        val mutableList1 = mutableListOf(1, 2, 3)
-        val mutableList2 = mutableListOf(0, 5, 7)
-        val listSet: MutableMultiSet<IntList> = mutableMultiSetOf(mutableList1, mutableList2)
+        val list1 = listOf(1, 2, 3)
+        val list2 = listOf(0, 5, 7)
+        val listSet: ConstMutableMultiSet<IntList> = constMutableMultiSetOf(list1, list2)
 
         expected = "[[1, 2, 3], [0, 5, 7]]"
-        assertEquals(expected, listSet.toString())
-
-        mutableList2.clear()
-        expected = "[[1, 2, 3], []]"
         assertEquals(expected, listSet.toString())
     }
 }
